Filtration.jl API

For the developers, here are the private methods.

Index

Filtration.Filtration

Filtration.Model

Filtration.Model

Filtration.Model

Filtration.Synthesis

Filtration.Synthesis

Filtration.DL_available_methods

Filtration.SL_available_methods

Filtration.check_root

Filtration.check_root

Filtration.solution

Filtration.solve

Documentation

Filtration.Filtration — Module

Filtration module.

Lists all the imported modules and packages:

Base

Core

DataInterpolations

DocStringExtensions

FiniteDifferences

ForwardDiff

LaTeXStrings

MINPACK

Nemo

OptimalControl

OrdinaryDiffEq

Plots

Roots

Symbolics

List of all the exported names:

DL_available_methods

Model

SL_available_methods

Synthesis

check_root

plot

solution

solve

Filtration.Model — Type

Type of a Model model, composed by:

e₊ : the function $e_+(x_1)$

e₋ : the function $e_-(x_1)$

f₊ : the function $f_+(x_1)$

f₋ : the function $f_-(x_1)$

g₊ : the function $g_+(x_1)$

g₋ : the function $g_-(x_1)$

criteria_type : the criteria type, either $:min$ or $:max$

λ₀ : the constant associated to the criteria type

dynamics! : the cost and state dynamics function

de₊ : the function $e_+'(x_1)$, compute thanks to ForwardDiff package

de₋ : the function $e_-'(x_1)$, compute thanks to ForwardDiff package

df₊ : the function $f_+'(x_1)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(x_1)$, compute thanks to ForwardDiff package

dg₊ : the function $g_+'(x_1)$, compute thanks to ForwardDiff package

dg₋ : the function $g_-'(x_1)$, compute thanks to ForwardDiff package

xₛ : the singular state $m_s$

uₛ : the singular control $u_s$

Δx₂ : the value $\Delta_x = x_f - x_2(T)$, where $T$ is the final time to leave the singular arc

Filtration.Model — Method

Model(
    e₊::Function,
    e₋::Function,
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function,
    criteria_type::Symbol;
    initial_guess
) -> Model

Constructor for Model structure.

Arguments

e₊ : the function $e_+(x₁)$

e₋ : the function $e_-(x₁)$

f₊ : the function $f_+(x₁)$

f₋ : the function $f_-(x₁)$

g₊ : the function $g_+(x₁)$

g₋ : the function $g_-(x₁)$

criteria_type : the criteria type, either $:min$ or $:max$

Keyword arguments (optional)

initialguess : initial guess for the singular state ``xs``, initialized to 1.

Returns

model : the Model model

Filtration.Model — Method

Model(
    e₊::Function,
    e₋::Function,
    f₊::Function,
    f₋::Function,
    criteria_type::Symbol;
    initial_guess
) -> Model

Constructor for Model structure.

Note

This function is a wrapper of Model function, where the functions $g_+$ and $g_-$ are set to 0 and 1 functions, respectively.

Arguments

e₊ : the function $e_+(x_1)$

e₋ : the function $e_-(x_1)$

f₊ : the function $f_+(x_1)$

f₋ : the function $f_-(x_1)$

criteria_type : the criteria type, either $:min$ or $:max$

Keyword arguments (optional)

initialguess : initial guess for the singular state ``xs``, initialized to 1.

Returns

model : the Model model

Filtration.Synthesis — Type

Type of a optimal synthesis, composed by:

model : a Model

tlim : the limits of time $t$

xlim : the limits of mass $m$

cut₋ : the cut₋ function

cut₊ : the cut₊ function

SL : the switching locus, as an ODESolution

DL : the dispersial locus, as Nothing if it not exists, and ODESolution if the method used to compute it is the continuation one, or a function if the method used is the direct one

Filtration.Synthesis — Method

Synthesis(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real};
    SL_method,
    DL_method
) -> Synthesis

Constructor for Model structure.

Arguments

model : a Model

tlim : the limits of time $t$

xlim : the limits of mass $m$

Keyword arguments (optional)

DL_method : method used to compute the dispersial locus curve, initialized to :continuation. See DL_available_methods for the list of available methods

Returns

plt : the Synthesis

Filtration.DL_available_methods — Method

DL_available_methods() -> Tuple{Symbol, Symbol}

Return the list of available methods to compute the dispersial locus function.

Filtration.SL_available_methods — Method

SL_available_methods() -> Tuple{Symbol, Symbol}

Return the list of available methods to compute the switching locus function.

Filtration.check_root — Method

check_root(
    e₊::Function,
    e₋::Function,
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function,
    criteria_type::Symbol;
    xlim
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the Model type.

Arguments

e₊ : the function $e_+(x_1)$

e₋ : the function $e_-(x_1)$

f₊ : the function $f_+(x_1)$

f₋ : the function $f_-(x_1)$

g₊ : the function $g_+(x_1)$

g₋ : the function $g_-(x_1)$

Keyword arguments (optional)

xlim : limits for x axis, initialized to (0,10)

Return

plt : the plot

Filtration.check_root — Method

check_root(
    e₊::Function,
    e₋::Function,
    f₊::Function,
    f₋::Function,
    criteria_type::Symbol;
    xlim
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the Model type.

Note

This function is a wrapper of check_root function, where the functions $g_+$ and $g_-$ are set to 0 and 1 functions, respectively.

Arguments

e₊ : the function $f_+(x_1)$

e₋ : the function $f_-(x_1)$

f₊ : the function $g_+(x_1)$

f₋ : the function $g_-(x_1)$

criteria_type : the criteria type, either $:min$ or $:max$

Keyword arguments (optional)

xlim : limits for x axis, initialized to (0,10)

Return

plt : the plot

Filtration.solution — Method

solution(
    synthesis::Synthesis,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Any

Wrapper to get solution from synthesis.

Arguments

model : a Model

initx₂ : the initial state ``x1(t_0)``

endx₂ : the final state ``x1(t_f)``

initx₁ : the initial state ``m0``

Returns

sol: the solution, as an ODESolution

Filtration.solve — Method

solve(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real;
    cut₋,
    cut₊,
    SL,
    DL,
    SL_end
) -> Any

Solve the optimal control problem associated to a Model model by indirect shooting.

Arguments

model : a Model

initx₂ : the initial state ``x1(t_0)``

endx₂ : the final state ``x1(t_f)``

initx₁ : the initial state ``m0``

Returns

sol: the solution, as an ODESolution

RecipesBase.plot — Method

plot(
    s::Synthesis;
    SL_nb_points,
    DL_nb_points,
    kwargs...
) -> Plots.Plot

Plot a Synthesis.

Arguments

synthesis : a Synthesis

Keyword arguments (optional)

DL_nb_points : number of points computed for the dispersial locus curve, initialized to 10

Returns

plt : the plot