Filtration.jl private functions

Index

Filtration.__compute_phi

Filtration.classification_plot

Filtration.get_arrow_time

Filtration.real_solve_control

Filtration.real_solve_singular

Filtration.synthesis_plot

Filtration.synthesis_plot

Documentation

Filtration.__compute_phi — Method

__compute_phi(
    t::Real,
    extremal::Extremal,
    model::Model;
    p
) -> Real

Compute the value of the function $t \mapsto \phi(x(t), p(t))$ on a given time $t$ of an Extremal computed on a Model.

Arguments

t : a real time $t$

extremal : the Extremal

model : the Model

Keyword arguments (optional)

p : a vector of parameter for the model (default is the parametes of the model).

Return

phi : the value of $\phi(t)$.

Filtration.classification_plot — Method

classification_plot(
    s::Synthesis;
    SL_nb_points,
    kwargs...
) -> Plots.Plot

Plot a Synthesis, in a classification of optimal control form

Arguments

synthesis : a Synthesis

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Returns

plt : the plot

Filtration.get_arrow_time — Method

get_arrow_time(
    value::Vector{<:Real},
    sol::SciMLBase.ODESolution,
    axis::Symbol
) -> Vector{<:Real}

Compute the time $t$ for which sol(t) ∈ y_value

Arguments

y_value : vector of y-axis value

sol : state trajectory as an ODESolution

Return

times : Vector of times

Filtration.real_solve_control — Method

real_solve_control(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real,
    N::Int64
) -> Solution

Transform a control extracted from a solution of a Model into a real control, composed only by -1 and 1 values. The singular arc is subdivided into N intervals, and the time passed with u =+1 is proportional to the singular control.

Arguments

model : a Model

control : a Control extracted from a solution of a Model

N : the number of subdivision of the singular arc

Return

new_control : a Control composed only by -1 and 1 values

Filtration.real_solve_singular — Method

real_solve_singular(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real,
    N::Int64;
    verbose
) -> Solution

Transform a control extracted from a solution of a Model into a real control, composed only by -1 and 1 values. The singular arc is subdivided into N intervals. On each subintervals, we ensure that the constol minimize the cost and saatisfy that the begining and the end of the state $x\_1$each subintervals corresponds to the value of the singular arc $x\_s$, and that the mean value of the state $x\_1$ on each subinterval is equal to the value of the singular arc $x\_s$.

Arguments

model : a Model

control : a Control extracted from a solution of a Model

N : the number of subdivision of the singular arc

Keyword arguments (optional)

verbose : if true, print the optimization process (default is false)

Return

new_control : a Control composed only by -1 and 1 values

Filtration.synthesis_plot — Method

synthesis_plot(
    multisynthesis::MultiSynthesis;
    kwargs...
) -> Plots.Plot

Plot a MultiSynthesis, in a feedback form.

Arguments

synthesis : a MultiSynthesis

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Returns

plt : the plot

Filtration.synthesis_plot — Method

synthesis_plot(
    s::Synthesis;
    SL_nb_points,
    kwargs...
) -> Plots.Plot

Plot a Synthesis, in a feedback form.

Arguments

synthesis : a Synthesis

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Returns

plt : the plot