Filtration.jl API

For the developers, here are the private methods.

Index

Filtration.Filtration

Filtration.MaxVolume

Filtration.cuts

Filtration.indirect_solve

Filtration.plot_synthesis

Filtration.switching_locus

Documentation

Filtration.Filtration — Module

Filtration module.

Lists all the imported modules and packages:

Base

Core

DocStringExtensions

ForwardDiff

LaTeXStrings

MINPACK

Nemo

OptimalControl

OrdinaryDiffEq

Plots

Roots

Symbolics

List of all the exported names:

MaxVolume

MinEnergy

cuts

indirect_solve

plot_synthesis

switching_locus

Filtration.MaxVolume — Type

Type of a Membrane Filtration Model, composed by:

f₊ : the function $f_+(m) = 0.5(f_1(m) + f_2(m))$

f₋ : the function $f_-(m) = 0.5(f_1(m) - f_2(m))$

g₊ : the function $g_+(m) = 0.5(g_1(m) + g_2(m))$

g₋ : the function $g_-(m) = 0.5(g_1(m) - g_2(m))$

dynamics : the cost and state dynamics function

dynamics! : the cost and state dynamics function

df₊ : the function $f_+'(m)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(m)$, compute thanks to ForwardDiff package

dg₊ : the function $g_+'(m)$, compute thanks to ForwardDiff package

dg₋ : the function $g_-'(m)$, compute thanks to ForwardDiff package

mₛ : the singular state $m_s$, compute as the root of $\psi$

uₛ : the singular control $u_s = - \frac{f_-(m_s)}{f_+(m_s)}$

λₛ : the singular costate $\lambda_s = -\frac{g(m_s)}{f_+(m_s)}$

Δt : the value $\Delta t = t_f - t_2$, where $t_2$ is the final time to leave the singular arc

Constructor Arguments

f₁ : the function $f_1(m)$

f₂ : the function $f_2(m)$

g₁ : the function $g₁(m)$

g₂ : the function $g₂(m)$

Filtration.cuts — Method

cuts(
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Tuple{Any, Any}

Compute the cuts $m_-(t)$ and $m_+(t)$, which corresponds to the curves which leads to $m_s$ at time $T$. The function $m₋$ is the solution of

$\dot m = f_-(m) - f_+(m), \quad m(T) = m_s,$

and the function $m₊$ is the solution of

$\dot m = f_-(m) + f_+(m), \quad m(T) = m_s,$

where $T = t_f - \Delta t$ corresponds to the optimal time to exit the singular arc and $m_s$ the singular state.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time

mlim : tuple of real which corresponds to the limits of mass

Returns

cut₋ : the function $m_-$, as an ODESolution

cut₊ : the function $m_+$, as an ODESolution

Filtration.indirect_solve — Method

indirect_solve(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to the membrane filtration model by indirect shooting.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial state $m(t_0)$

Returns

sol: the solution, as an ODESolution

Filtration.plot_synthesis — Method

plot_synthesis(
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Plots.Plot

Plot the optimal synthesis of a MaxVolume.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time

mlim : tuple of real which corresponds to the limits of mass

Returns

plt : the plot

Filtration.switching_locus — Method

switching_locus(
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Any

Compute the function $S(m)$ which give the time with respect to the deposit mass $m$ which belongs to the switching locus. This function is the solution of

$S'(m) = \frac{1}{f_+(m) + f_-(m)} - \frac{ \gamma'(m)}{\tilde g'(m_T(m)) (f_+(m_T(m)) + f_-(m_T(m)))}, \quad S(mₛ) = T$

where $m_T(m) = \tilde g^{-1}(\gamma(m))$, $\tilde g(m) = g_-(m) + g_+(m)$, $T = t_f - \Delta t$ and

$\gamma(m) = - \frac{f_-(m) g_+(m)}{f_+(m)} + g_-(m)$

The value of $\tilde g^{-1}(y)$ is numerically comute by solving $\tilde g(x) - y = 0$.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time

mlim : tuple of real which corresponds to the limits of mass

Returns

switching_locus : the function $S(m)$, as an ODESolution