Filtration.jl private functions

Index

Filtration.cuts

Filtration.dispersal_locus

Filtration.dispersal_locus_continuation

Filtration.dispersal_locus_direct

Filtration.dispersal_locus_direct_function

Filtration.end_switching_locus_direct

Filtration.solve

Filtration.solve_MSP

Filtration.solve_P

Filtration.solve_PSP

Filtration.switching_locus

Filtration.switching_locus_continuation

Filtration.switching_locus_direct

Filtration.switching_locus_direct_function

Documentation

Filtration.cuts — Method

cuts(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> Tuple{SciMLBase.ODESolution, SciMLBase.ODESolution}

Compute the cuts $c_-(t)$ and $c_+(t)$, which corresponds to the curves which leads to the singular state $x_s$ at time $T$. The function $c₋$ is the solution on $[t_0, T]$ of

\[\left\{ \begin{array}{ll} \dot x_0 = e_-(x_1) - e_+(x_1), & \quad x_0(T) = 0, \\\dot x_1 = f_-(x_1) - f_+(x_1), & \quad x_1(T) = x_s, \\\dot x_2 = g_-(x_1) - g_+(x_1), & \quad x_2(T) = x_f- \Delta_x, \end{array} \right.\]

and the function $m₊$ is the solution on $[T, t_f]$ of

\[\left\{\begin{array}{ll} \dot x_0 = e_-(x_1) + e_+(x_1), & \quad x_O(T) = 0, \\\dot x_1 = f_-(x_1) + f_+(x_1), & \quad x_1(T) = x_s, \\\dot x_2 = g_-(x_1) + g_+(x_1), & \quad x_2(T) = x_f - \Delta_x, \end{array} \right.\]

Warning

The cost $x_0(T)$ at time $T$ is initialized to 0.

Arguments

model : a Model

xlim : the limits of state $x_2$

ylim : the limits of state $x_1$

Returns

cut₋ : the function $c_-$

cut₊ : the function $c_+$

Filtration.dispersal_locus — Method

dispersal_locus(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real},
    SL::Union{Function, SciMLBase.ODESolution},
    SL_end::Real;
    method
) -> Union{Filtration.var"#DL#67"{Model}, SciMLBase.ODESolution}

Compute the dispersal locus function, with respect to the method given in argument.

Arguments

model : a Model

xlim : the limits of state $x_2$

ylim : the limits of state $x_1$

SL : the switching locus, as an ODESolution

SL_end : the end of the switching locus (on $x_2$)

Keyword arguments (optional)

method : the method to compute the dispersal locus, initialized to :continuation. See available_methods for the list of available methods

Returns

DL : the dispersal locus function

Filtration.dispersal_locus_continuation — Method

dispersal_locus_continuation(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real},
    SL::Union{Function, SciMLBase.ODESolution},
    SL_end::Real
) -> SciMLBase.ODESolution

dispersal locus function computed by continuation method.

A state $[x_1, x_2]$ belongs to the dispersal curve if $S(x_2, x_1) = 0$, where $S \colon \mathbb R^2 \to \mathbb R$ is defined by

\[S(x_1, x_2) = \varphi_0^+(x_1, x_2) - \varphi_0^{-+}(x_1, x_2)\]

where $\varphi_0^+$ correponds to the final cost with the control u₊, and where $\varphi_0^{-+}$ the final cost with the control u₋ ∘ u₊.

We suppose that there exsits the switching locus $DL(x_1)$ such that $S(x_1, DL(x_1)) = 0$. Since $S$ is constant, we have

\[ \frac{\partial S}{\partial x_1}(x_1, DL(x_1)) + \frac{\partial S}{\partial x_2}(x_1, DL(x_1)) DL'(x_1) = 0\]

By the previous hypothesis, $\frac{\partial S}{\partial x_2}(x_1, DL(x_1))$ is invertible. We have so

\[DL'(x_1) = -\left( \frac{\partial S}{\partial x_2}(x_1, DL(x_1)) \right)^{-1} \frac{\partial S}{\partial x_1}(x_1, DL(x_1)).\]

The continuation method consists to solve this ODE with the initial condition $DL(x_s) = x_f - \Delta x_2$. Thanks to the combinaison of ForwardDiff and OrdinaryDiffEq package, the derivative of $\varphi_0^{+}$ and $\varphi_0^{-+}$ are computed by automatic differentiation methods by using the variational equations.

Warning

This function must be used only when a dispersal locus happened, ie. when the end of the switching curve (given by the end_switching_locus function) is below ylim[2].

Filtration.dispersal_locus_direct — Method

dispersal_locus_direct(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real},
    SL::Union{Function, SciMLBase.ODESolution},
    SL_end::Real
) -> Filtration.var"#DL#67"{Model, var"#s182", var"#s1821", _A, <:Real} where {var"#s182"<:Tuple{Real, Real}, var"#s1821"<:Tuple{Real, Real}, _A}

Construct the dispersal locus function by direct method.

Arguments

model : a Model

xlim : the limits of state $x_2$

ylim : the limits of state $x_1$

SL : the switching locus, as an ODESolution

SL_end : the end of the switching locus (on $x_2$)

Returns

DL : the dispersal locus function

Filtration.dispersal_locus_direct_function — Method

dispersal_locus_direct_function(
    m::Real,
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real},
    SL::Union{Function, SciMLBase.ODESolution},
    SL_end::Real
) -> Real

dispersal locus function by direct method, which corresponds to the curve where the cost associated to u₋ ∘ u₊ is equal to the cost associated to u₊. This function is parametrized like $DL(x_1) = x_2$

From $x_1$ fixed, we find a zero of the function $S \colon \mathbb R \to \mathbb R$ defined by $S(x_2) = \varphi_0^+(x_1, x_2) - \varphi_0^{-+}(x_1, x_2)$ where $\varphi_0^+$ correponds to the final cost with the control u₊, and where $\varphi_0^{-+}$ the final cost with the control u₋ ∘ u₊.

Warning

This function must be used only when a dispersal locus happened, ie. when the end of the switching curve (given by the end_switching_locus function) is below ylim[2].

Arguments

x₁ : the state $x_₁$

model : a Model

xlim : the limits of state $x_2$

ylim : the limits of state $x_1$

SL : the switching locus, as an ODESolution

SL_end : the end of the switching locus

Return

x₂ : the state $x_2$

Filtration.end_switching_locus_direct — Method

end_switching_locus_direct(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> Real

End of the switching locus function $SL(x_1) = x_2$ computed by direct method

This end point $x_1$ is given by finding a root of

\[ SL'(x_1) * (f₋(x_1) - f₊(x_1)) - (g₋(x_1) - g₊(x_1))\]

Since a rootfinding function is used in the computation of the switching locus function, the derivative of $SL$ is computed by finite differences method, thanks to the FiniteDifferences package.

Note

If no end of switching locus is found, the value return is the maximum limits of y, ie. ylim[2].

Arguments

model: a Model

xlim : the limits of the state $x_2$

ylim : the limits of the state $x_1$

Keyword arguments (optional)

initial_guess : initial guess for $x_1$, initialized to the singular state $x_s$ if not provided

Return

x₁ : the end of the switching locus

Filtration.solve — Method

solve(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real;
    cut₋,
    cut₊,
    SL,
    DL,
    SL_end
) -> SciMLBase.ODESolution

Solve the optimal control problem associated to a Model model by indirect shooting.

Arguments

model : a Model

init_x₂ : the initial state $x_1(t_0)$

end_x₂ : the final state $x_1(t_f)$

init_x₁ : the initial state $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.solve_MSP — Method

solve_MSP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real;
    end_SL
) -> SciMLBase.ODESolution

Solve the optimal control problem associated to a Model model, with the structure given by $u_- \circ u_s \circ u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_1(t_0)$

end_x₂ : the final state $x_1(t_f)$

init_x₁ : the initial state $x_2(t_0)$

Returns

sol: the solution, as an ODESolution

Filtration.solve_P — Method

solve_P(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> SciMLBase.ODESolution

Solve the optimal control problem associated to a Model model, with the structure given by $u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_1(t_0)$

end_x₂ : the final state $x_1(t_f)$

init_x₁ : the initial state $x_2(t_0)$

Returns

sol: the solution, as an ODESolution

Filtration.solve_PSP — Method

solve_PSP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> SciMLBase.ODESolution

Solve the optimal control problem associated to the membrane filtration model by indirect shooting, with the structure given by $u_+ \circ u_s \circ u_+$

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_1(t_0)$

end_x₂ : the final state $x_1(t_f)$

init_x₁ : the initial state $x_2(t_0)$

Returns

sol: the solution, as an ODESolution

Filtration.switching_locus — Method

switching_locus(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real};
    method,
    initial_guess
) -> Union{Filtration.var"#SL#108"{<:Real, Model, <:Tuple{Real, Real}}, SciMLBase.ODESolution}

Provide the switchig locus with respect to the method. To see all method available, please use available_methods function.

Arguments

model: a Model

xlim : limits for state $x_2$

ylim : limits for state $x_1$

Keyword arguments (optional)

method : method used to compute the switching locus, initialized to :continuation

initial_guess : initial guess for $x_{1,f}$, initialized to the singular state $x_s$ if not provided, only used if method = :direct

Return

SL : the switching locus

Filtration.switching_locus_continuation — Method

switching_locus_continuation(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> SciMLBase.ODESolution

Switching locus function computed by continuation method.

A state $[x_1, x_2]$ belongs to the switching curve if $S(x_2, x_1) = 0$, where $S \colon \mathbb R^2 \to \mathbb R$ is defined by

\[S(x_2, x_1) = H(x_1, \varphi_1(x_1, x_2))\]

The function $\varphi_1$ coresponds to the solution at the final time of the state $x_1$ by applying the control $u_+$, and the function $H$ is defined by

\[ H(x_1, x_{1,f}) = \lambda_1(x_1, x_{1,f}) (f_+(x_1) + f_-(x_1)) + \lambda_2(x_{1,f})(g_+(x_1) + g_-(x_1)) - (e_+(x_1) + e_-(x_1))\]

where

\[ \lambda_1(x_1, x_{1,f}) = \frac{e_+(x_1) - \lambda_2(x_{1,f}) g_+(x_1)}{f_+(x_1)}\]

and

\[ \lambda_2(x_{1,f}) = \frac{e_+(x_1) + e_-(x_1)}{g_+(x_1) + g_-(x_1)}. \]

We suppose that there exsits the switching locus $SL(x_1)$ such that $S(x_1, SL(x_1)) = 0$. Since $S$ is constant, we have

\[ \frac{\partial S}{\partial x_1}(x_1, SL(x_1)) + \frac{\partial S}{\partial x_2}(x_1, SL(x_1)) SL'(x_1) = 0\]

By the previous hypothesis, $\frac{\partial S}{\partial x_2}(x_1, SL(x_1))$ is invertible. We have so

\[SL'(x_1) = -\left( \frac{\partial S}{\partial x_2}(x_1, SL(x_1)) \right)^{-1} \frac{\partial S}{\partial x_1}(x_1, SL(x_1)).\]

The continuation method consists to solve this ODE with the initial condition $SL(x_s) = x_f - \Delta x_2$. Thanks to the combinaison of ForwardDiff and OrdinaryDiffEq package, the derivative of $\varphi_1$ is computed by automatic differentiation methods by using the variatomal equations.

Arguments

model: a Model

xlim : the limits of the state $x_2$

ylim : the limits of the state $x_1$

Return

SL : the switching locus function, as an ODESolution

Filtration.switching_locus_direct — Method

switching_locus_direct(
    model::Model,
    xlim::Tuple{Real, Real};
    initial_guess
) -> Filtration.var"#SL#108"{<:Real, Model, <:Tuple{Real, Real}}

Wrapper for the switching locus by direct method, computed by switching_locus_direct_function.

Arguments

model : a Model

xlim : the limits of the state $x_2$

Keyword arguments (optional)

initial_guess : initial guess for $x_{1,f}$, initialized to the singular state $x_s$ if not provided

Retrun

SL : the switching locus

Filtration.switching_locus_direct_function — Method

switching_locus_direct_function(
    x₁::Real,
    model::Model,
    xlim::Tuple{Real, Real};
    initial_guess
) -> Real

Switching locus function $SL(x_1) = x_2$ computed by direct method

From a given state $x_1$ which belongs to this switching curve, the associated final state $x_{1,f}$ is computed by finding a zero of

\[ H(x_1, x_{1,f}) = \lambda_1(x_1, x_{1,f}) (f_+(x_1) + f_-(x_1)) + \lambda_2(x_{1,f})(g_+(x_1) + g_-(x_1)) - (e_+(x_1) + e_-(x_1))\]

where

\[ \lambda_1(x_1, x_{1,f}) = \frac{e_+(x_1) - \lambda_2(x_{1,f}) g_+(x_1)}{f_+(x_1)}\]

and

\[ \lambda_2(x_{1,f}) = \frac{e_+(x_1) + e_-(x_1)}{g_+(x_1) + g_-(x_1)}. \]

From the final state $x(t_f) = [x_{1,f}, x_f]$, the state $x(T)$ is computed by a backward integration of the dynamic which stop when $x_1(t) = x_1$, thanks to a ContinuousCallback.

Warning

The given state $x_1$ must belong to the switching locus curve, ie. $x_s \leq x_1 \leq x_d$ where $x_s$ is the singular state, given by model.xₛ, and where $x_d$ is the end of the dispersal locus, computed by the end_switching_locus function.

Arguments

x₁ : the state $x_1$

model : a Model

xlim : the limits of the state $x_2$

Keyword arguments (optional)

initial_guess : initial guess for $x_{1,f}$, initialized to the singular state $x_s$ if not provided

Retrun

x₂ : the state $x_2$