Filtration.jl API

For the developers, here are the private methods.

Index

Filtration.Filtration

Filtration.MembraneFiltrationModel

Filtration.cuts

Filtration.indirect_solve

Filtration.optimal_synthesis

Filtration.switching_locus

Documentation

Filtration.Filtration — Module

Filtration module.

Lists all the imported modules and packages:

Base

Core

DocStringExtensions

ForwardDiff

LaTeXStrings

MINPACK

Nemo

OptimalControl

OrdinaryDiffEq

Plots

Roots

Symbolics

List of all the exported names:

MembraneFiltrationModel

cuts

indirect_solve

optimal_synthesis

switching_locus

Filtration.MembraneFiltrationModel — Type

Type of a Membrane Filtration Model, composed by:

f₁ : the function $f_1(m)$

f₂ : the function $f_2(m)$

g : the function $g(m)$

f₊ : the function $f_+(m) = 0.5(f_1(m) + f_2(m))$

f₋ : the function $f_-(m) = 0.5(f_1(m) - f_2(m))$

dg : the function $g'(m)$, compute thanks to ForwardDiff package

df₊ : the function $f_+'(m)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(m)$, compute thanks to ForwardDiff package

state_dynamic : the state dynamic function, given by $f(x,u) = f_-(m) + u f_+(m)$

cost_dynamic : the cost dynamic function, given by $f^0(x,u) = u g(m)$

dynamics! : the cost and state dynamics

ψ : the function $\psi(m) = g(m) [f'(m) f_+(m) - f_-(m) f'(m)] + g'(m) f_+(m) f_-(m)$

γ : the function $\gamma(m) = - \frac{g(m) f_-(m)}{f_+(m)}$

Φ : the function $\phi(m, \lambda) = \lambda f_+(m) + g(m)$

mₛ : the singular state $m_s$, compute as the root of $\psi$

uₛ : the singular control $u_s = - \frac{f_-(m_s)}{f_+(m_s)}$

λₛ : the singular costate $\lambda_s = -\frac{g(m_s)}{f_+(m_s)}$

dynamics₋! : the cost and state dynamics, with $u=-1$

dynamics₊! : the cost and state dynamics, with $u=+1$

dynamicsₛ! : the cost and state dynamics, with $u=u_s$

Δt : the value $\Delta t = t_f - t_2$, where $t_2$ is the final time to leave the singular arc

Constructor Arguments

f₁ : the function $f_1(m)$

f₂ : the function $f_2(m)$

g : the function $g(m)$

During the construction, $f_1$ and $g$ are tested to be L-functions, and $f_2$ is tested to be a K-function.

Filtration.cuts — Method

cuts(
    model::MembraneFiltrationModel,
    t0::Real,
    tf::Real
) -> Tuple{Any, Any}

Compute the cuts $m_-(t)$ and $m_+(t)$, which corresponds to the curves which leads to $m_s$ at time $t_2$. The function $m₋$ is the solution of

$\dot m = f_-(m) - f_+(m), \quad m(t_2) = m_s,$

and the function $m₊$ is the solution of

$\dot m = f_-(m) + f_+(m), \quad m(t_2) = m_s,$

where $t_2 = t_f - \Delta t$ corresponds to the optimal time to exit the singular arc and $m_s$ the singular state.

Arguments

model : a MembraneFiltrationModel

t0 : the initial time $t_0$

tf : the final time $t_f$

Returns

m₋ : the function $m_-$, as an ODESolution

m₊ : the function $m_+$, as an ODESolution

Filtration.indirect_solve — Method

indirect_solve(
    model::MembraneFiltrationModel,
    t0::Real,
    tf::Real,
    m0::Real;
    ylim
) -> Any

Solve the optimal control problem associated to the membrane filtration model by indirect shooting.

Arguments

model : a MembraneFiltrationModel

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial state $m(t_0)$

ylim : the minimum and maximum limits of state, initialized to (0,10)

Returns

sol: the solution, as an ODESolution

Filtration.optimal_synthesis — Method

optimal_synthesis(
    model::MembraneFiltrationModel,
    t0::Real,
    tf::Real;
    ylim
) -> Plots.Plot

Plot the optimal synthesis of a MembraneFiltrationModel.

Arguments

model : a MembraneFiltrationModel

t0 : the initial time $t_0$

tf : the final time $t_f$

ylim : the limits of the y-axis, initialized to (0, 10)

Returns

plt : the plot

Filtration.switching_locus — Method

switching_locus(
    model::MembraneFiltrationModel,
    tf::Real;
    ylim
) -> Any

Compute the function $T(m)$ which give the time with respect to the deposit mass $m$ which belongs to the switching locus. This function is the solution of

$T'(m) = \frac{1}{f_1(m)} - \frac{ \gamma'(m)}{g'(m_T(m))f_1(m_T(m))}, \quad T(mₛ) = t_2$

where $m_T(m) = g^{-1}(\gamma(m))$ and $t_2 = t_f - \Delta t$. The value of $g^{-1}(y)$ is numerically comute by solving $g(x) - y = 0$.

Arguments

model : a MembraneFiltrationModel

tf : the final time $t_f$

ylim : the limits of the y-axis, initialized to (0, 10)

Returns

T : the function $T(m)$, as an ODESolution