Filtration.jl API

For the developers, here are the private methods.

Index

Filtration.Filtration

Filtration.MaxVolume

Filtration.MaxVolume

Filtration.MinEnergy

Filtration.MinEnergy

Filtration.cuts

Filtration.cuts

Filtration.dispersial_locus

Filtration.dispersial_locus

Filtration.end_switching_locus

Filtration.indirect_solve

Filtration.indirect_solve

Filtration.indirect_solve_MP

Filtration.indirect_solve_MP

Filtration.indirect_solve_MSP

Filtration.indirect_solve_MSP

Filtration.indirect_solve_P

Filtration.indirect_solve_P

Filtration.indirect_solve_PSP

Filtration.indirect_solve_PSP

Filtration.plot_max_volume

Filtration.plot_min_energy

Filtration.plot_synthesis

Filtration.plot_synthesis

Filtration.switching_locus

Filtration.switching_locus

Documentation

Filtration.Filtration — Module

Filtration module.

Lists all the imported modules and packages:

Base

Core

DataInterpolations

DocStringExtensions

FiniteDifferences

ForwardDiff

LaTeXStrings

MINPACK

Nemo

OptimalControl

OrdinaryDiffEq

Plots

Roots

Symbolics

List of all the exported names:

MaxVolume

MinEnergy

cuts

dispersial_locus

end_switching_locus

indirect_solve

indirect_solve_MP

indirect_solve_MSP

indirect_solve_P

indirect_solve_PSP

plot_max_volume

plot_min_energy

plot_synthesis

switching_locus

Filtration.MaxVolume — Type

Type of a MaxVolume model, composed by:

f₊ : the function $f_+(m)$

f₋ : the function $f_-(m)$

g₊ : the function $g_+(m)$

g₋ : the function $g_-(m)$

dynamics : the cost and state dynamics function

dynamics! : the cost and state dynamics function

df₊ : the function $f_+'(m)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(m)$, compute thanks to ForwardDiff package

dg₊ : the function $g_+'(m)$, compute thanks to ForwardDiff package

dg₋ : the function $g_-'(m)$, compute thanks to ForwardDiff package

mₛ : the singular state $m_s$

uₛ : the singular control $u_s$

λₛ : the singular costate $\lambda_s$

Δt : the value $\Delta t = t_f - T$, where $T$ is the final time to leave the singular arc

Filtration.MaxVolume — Method

MaxVolume(
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function;
    mₛ_guess
) -> MaxVolume

Constructor for MaxVolume structure.

Arguments

f₊ : the function $f_+(m)$

f₋ : the function $f_-(m)$

g₊ : the function $g_+(m)$

g₋ : the function $g_-(m)$

mₛguess : initial guess for the singular state ``ms``, initialized to 0.5

Returns

model : the MaxVolume model

Filtration.MinEnergy — Type

Type of a MinEnergy model, composed by:

f₊ : the function $f_+(m)$

f₋ : the function $f_-(m)$

g₊ : the function $g_+(m)$

g₋ : the function $g_-(m)$

e₊ : the function $e_+(m)$

e₋ : the function $e_-(m)$

dynamics : the cost and state dynamics function

dynamics! : the cost and state dynamics function

df₊ : the function $f_+'(m)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(m)$, compute thanks to ForwardDiff package

dg₊ : the function $g_+'(m)$, compute thanks to ForwardDiff package

dg₋ : the function $g_-'(m)$, compute thanks to ForwardDiff package

de₊ : the function $e_+'(m)$, compute thanks to ForwardDiff package

de₋ : the function $e_-'(m)$, compute thanks to ForwardDiff package

mₛ : the singular state $m_s$

uₛ : the singular control $u_s$

λₘₛ : the singular costate $\lambda_{ms}$ associated to the state $m$

λᵥ : the constant costate $\lambda_v$ associated to the state $v$

Δv : the value $\Delta_v = v_f - v(T)$, where $T$ is the final time to leave the singular arc

Δt : the value $\Delta t = t_f - T$, where $T$ is the final time to leave the singular arc

Filtration.MinEnergy — Method

MinEnergy(
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function,
    e₊::Function,
    e₋::Function;
    mₛ_guess
) -> MinEnergy

Constructor for MinEnergy structure.

Arguments

f₊ : the function $f_+(m)$

f₋ : the function $f_-(m)$

g₊ : the function $g_+(m)$

g₋ : the function $g_-(m)$

e₊ : the function $e_+(m)$

e₋ : the function $e_-(m)$

Returns

model : the MinEnergy model

Filtration.cuts — Method

cuts(
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Tuple{Any, Any}

Compute the cuts $m_-(t)$ and $m_+(t)$, which corresponds to the curves which leads to $m_s$ at time $T$. The function $m₋$ is the solution on $[0, T]$ of

\[\left\{\begin{array}{ll} \dot v = g_-(m) - g_+(m), & \quad v(T) = 0, \\\dot m = f_-(m) - f_+(m), & \quad m(T) = m_s, \end{array} \right.\]

and the function $m₊$ is the solution on $[0, T]$ of

\[\left\{\begin{array}{ll} \dot v = g_-(m) + g_+(m), & \quad v(T) = 0 , \\\dot m = f_-(m) + f_+(m), & \quad m(T) = m_s, \end{array} \right.\]

where $T = t_f - \Delta t$ corresponds to the optimal time to exit the singular arc and $m_s$ the singular state.

Warning

The cost $v(T)$ at the initial time $T$ is initialized to 0.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time $t$

mlim : tuple of real which corresponds to the limits of mass $m$

Returns

cut₋ : the function $m_-$, as an ODESolution

cut₊ : the function $m_+$, as an ODESolution

Filtration.cuts — Method

cuts(
    model::MinEnergy,
    vlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Tuple{Any, Any}

Compute the cuts $m_-(t)$ and $m_+(t)$, which corresponds to the curves which leads to $m_s$ at time $T$. The function $m₋$ is the solution on $[T, tf]$ of

\[\left\{ \begin{array}{ll} \dot e = e_-(m) - e_+(m), & \quad e(T) = 0, \\\dot m = f_-(m) - f_+(m), & \quad m(T) = m_s, \\\dot v = g_-(m) - g_+(m), & \quad v(T) = v_T, \end{array} \right.\]

and the function $m₊$ is the solution on $[T, tf]$ of

\[\left\{\begin{array}{ll} \dot e = e_-(m) + e_+(m), & \quad e(T) = 0, \\\dot m = f_-(m) + f_+(m), & \quad m(T) = m_s, \\\dot v = g_-(m) + g_+(m), & \quad v(T) = v_T, \end{array} \right.\]

where $T = t_f - \Delta t$ corresponds to the optimal time to exit the singular arc, $v_T = v_f - \Delta_v$ the associated state $v$ and $m_s$ the singular state.

Warning

The cost $e(T)$ at time $T$ is initialized to 0.

Arguments

model : a MinEnergy

vlim : tuple of real which corresponds to the limits of volume $v$

mlim : tuple of real which corresponds to the limits of mass $m$

Returns

cut₋ : the function $m_-$, as an ODESolution

cut₊ : the function $m_+$, as an ODESolution

Filtration.dispersial_locus — Method

dispersial_locus(
    t::Real,
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real};
    m_guess
) -> Float64

Compute some points of the dispersial locus curve, which corresponds to the curve where the cost associated to u₋ ∘ u₊ is equal to the cost associated to u₊.

Warning

This function must be used only when a dispersial locus happened, ie. when the switching locus don't hit the bounds.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time $t$

mlim : tuple of real which corresponds to the limits of mass $m$

SL : an ODESolution which corresponds to the switching locus computed by switching_locus function

Returns

tspan_DL : array of time t

mspan_DL : array of state m

Filtration.dispersial_locus — Method

dispersial_locus(
    m::Real,
    model::MinEnergy,
    vlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Any

Dispersial locus function, which gives the volume $v$ from a mass $m$, which corresponds to the curve where the cost associated to u₋ ∘ u₊ is equal to the cost associated to u₊.

Warning

This function must be used only when a dispersial locus happened, ie. when the end of the switching curve (given by the end_switching_locus function) is below the maximal bound of $m$.

Arguments

m : the mass $m$

model : a MinEnergy

vlim : tuple of real which corresponds to the limits of volume $v$

Return

v : the volume $v$

Filtration.end_switching_locus — Method

end_switching_locus(
    model::MinEnergy,
    vlim::Tuple{Real, Real};
    m_guess
) -> Any

Compute the mass $m$ which corresponds to the end of the switching locus.

Denoting $SL$ the switching locus function, this end point $m$ is given by finding a root of

\[ SL'(m) * (f₋(m) - f₊(m)) - (g₋(m) - g₊(m))\]

Since a rootfinding function is used in the computation of the switching locus function, the derivative of $SL$ is computed by finite differences method, thanks to the FiniteDifferences package.

Arguments

model: a MinEnergy

vlim : tuple of real which corresponds to the limits of volume $v$

mguess : initial guess for $m$, initialized to the singular state ``ms`` if not provided

Return

m : the limit mass $m$

Filtration.indirect_solve — Method

indirect_solve(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MaxVolume model by indirect shooting.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve — Method

indirect_solve(
    model::MinEnergy,
    v0::Real,
    vf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MinEnergy model by indirect shooting.

Arguments

model : a MinEnergy

v0 : the initial volume $v_0$

vf : the final volume $v_f$

m0 : the initial state $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_MP — Method

indirect_solve_MP(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MaxVolume model by indirect method, with the structure given by $u_- \circ u_+$, and when the change is done by intersection with the switching curve

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_MP — Method

indirect_solve_MP(
    model::MinEnergy,
    v0::Real,
    vf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MinEnergy model by indirect method, with the structure given by $u_- \circ u_+$, and when the change is done by intersection with the switching curve

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MinEnergy

v0 : the initial volume $v_0$

vf : the final volume $v_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_MSP — Method

indirect_solve_MSP(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MaxVolume model, with the structure given by $u_- \circ u_s \circ u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_MSP — Method

indirect_solve_MSP(
    model::MinEnergy,
    v0::Real,
    vf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MinEnergy model, with the structure given by $u_- \circ u_s \circ u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MinEnergy

v0 : the initial volume $v_0$

vf : the final volume $v_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_P — Method

indirect_solve_P(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MaxVolume model, with the structure given by $u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_P — Method

indirect_solve_P(
    model::MinEnergy,
    v0::Real,
    vf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MinEnergy model, with the structure given by $u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MinEnergy

v0 : the initial volume $v_0$

vf : the final volume $v_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_PSP — Method

indirect_solve_PSP(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to a MaxVolume model by indirect shooting, with the structure given by $u_+ \circ u_s \circ u_+$

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_PSP — Method

indirect_solve_PSP(
    model::MinEnergy,
    v0::Real,
    vf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to the membrane filtration model by indirect shooting, with the structure given by $u_+ \circ u_s \circ u_+$

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use indirect_solve function instead.

Arguments

model : a MinEnergy

v0 : the initial volume $v_0$

vf : the final volume $v_f$

m0 : the initial mass $m_0$

Returns

sol: the solution, as an ODESolution

Filtration.plot_max_volume — Function

plot_max_volume(
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function
) -> Plots.Plot
plot_max_volume(
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function,
    xlim::Tuple{Real, Real}
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the MaxVolume type.

Arguments

f₊ : the function $f_+(m)$

f₋ : the function $f_-(m)$

g₊ : the function $g_+(m)$

g₋ : the function $g_-(m)$

xlim : limits for x axis, initialized to (0,10)

Return

plt : the plot

Filtration.plot_min_energy — Method

plot_min_energy(
    f₊::Function,
    f₋::Function,
    g₊::Function,
    g₋::Function,
    e₊::Function,
    e₋::Function;
    xlim
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the MinEnergy type.

Arguments

f₊ : the function $f_+(m)$

f₋ : the function $f_-(m)$

g₊ : the function $g_+(m)$

g₋ : the function $g_-(m)$

e₊ : the function $e_+(m)$

e₋ : the function $e_-(m)$

xlim : limits for x axis, initialized to (0,10)

Return

plt : the plot

Filtration.plot_synthesis — Method

plot_synthesis(
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real};
    DL_nb_points
) -> Plots.Plot

Plot the optimal synthesis of a MaxVolume.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time $t$

mlim : tuple of real which corresponds to the limits of mass $m$

DLnbpoints : number of points computed for the dispersial locus curve, initialized to 10

Returns

plt : the plot

Filtration.plot_synthesis — Method

plot_synthesis(
    model::MinEnergy,
    vlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real};
    SL_nb_points,
    DL_nb_points
) -> Plots.Plot

Plot the optimal synthesis of a MinEnergy.

Arguments

model : a MinEnergy

vlim : tuple of real which corresponds to the limits of time $v$

mlim : tuple of real which corresponds to the limits of mass $m$

SLnbpoints : number of points computed for the switching locus curve, initialized to 10

DLnbpoints : number of points computed for the dispersial locus curve, initialized to 10

Returns

plt : the plot

Filtration.switching_locus — Method

switching_locus(
    model::MaxVolume,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Any

Compute the switching locus $S(m)$ which give the time with respect to the deposit mass $m$. This function is the solution of

\[S'(m) = \frac{1}{f_+(m) + f_-(m)} - \frac{ \gamma'(m)}{\tilde g'(m_T(m)) (f_+(m_T(m)) + f_-(m_T(m)))}, \quad S(mₛ) = T \]

where

\[\begin{aligned} m_T(m) &= \tilde g^{-1}(\gamma(m)), \\\tilde g(m) &= g_-(m) + g_+(m), \\\gamma(m) &= - \frac{f_-(m) g_+(m)}{f_+(m)} + g_-(m), \end{aligned}\]

and $T = t_f - \Delta t$. The value of $\tilde g^{-1}(y)$ is numerically comute by solving $\tilde g(x) - y = 0$.

In order to only compute the switching locus and not the dispersial one, the integration of the previous ODE stop when

\[S'(m) (f_-(m) - f_+(m)) - 1 = 0. \]

This condition can be handle thanks to the ContinuousCallback function of the OrdinaryDiffEq package.

Arguments

model : a MaxVolume

tlim : tuple of real which corresponds to the limits of time $t$

mlim : tuple of real which corresponds to the limits of mass $m$

Returns

switching_locus : the function $S(m)$, as an ODESolution

Filtration.switching_locus — Method

switching_locus(
    m::Real,
    model::MinEnergy,
    vlim::Tuple{Real, Real};
    m_guess
) -> Any

Switching locus function, which gives the volume $v$ from the mass $m$.

From a given state $m$ which belongs to this switching curve, the associated final state $m_f$ is computed by finding a zero of

\[ H(m, m_f) = \lambda_m(m, m_f) (f_+(m) + f_-(m)) + \lambda_v(m_f)(g_+(m) + g_-(m)) - (e_+(m) + e_-(m))\]

where

\[ \lambda_m(m, m_f) = \frac{e_+(m) - \lambda_v(m_f) g_+(m)}{f_+(m)}\]

and

\[ \lambda_v(m_f) = \frac{e_+(m) + e_-(m)}{g_+(m) + g_-(m)}. \]

From this final state $m_f$, with the final volume $v_f$, the volume $v$ is computed by a backward integration of the dynamic which stop when $m(t) = m$, thanks to a ContinuousCallback.

Warning

The given state $m$ must belong to the switching locus curve, ie. $m_s \leq m \leq m_d$ where $m_s$ is the singular state, given by model.mₛ, and where $m_d$ is the end of the dispersial locus, computed by the end_switching_locus function.

Arguments

m : the mass $m$

model : a MinEnergy

vlim : tuple of real which corresponds to the limits of volume $v$

mguess : initial guess for $m$, initialized to the singular state ``ms`` if not provided

Retrun

v : the volume $v$