Filtration.jl private functions

Index

Filtration.__compute_phi

Filtration.__get_time

Filtration.__sep

Filtration.classification_plot

Filtration.dispersal_locus_2

Filtration.get_arrow_time

Filtration.real_solve_control

Filtration.real_solve_optimal

Filtration.real_solve_singular

Filtration.synthesis_plot

Filtration.synthesis_plot

Documentation

Filtration.__compute_phi — Method

__compute_phi(
    t::Real,
    extremal::Extremal,
    model::Model;
    p
) -> Real

Compute the value of the function $t \mapsto \phi(x(t), p(t))$ on a given time $t$ of an Extremal computed on a Model.

Arguments

t : a real time $t$

extremal : the Extremal

model : the Model

Keyword arguments (optional)

p : a vector of parameter for the model (default is the parametes of the model).

Return

phi : the value of $\phi(t)$.

Filtration.__get_time — Method

__get_time(
    synthesis1::Synthesis,
    synthesis2::Synthesis,
    xmax::Real
) -> Real

Provide the time $t_2$ on the singuar arc associated to synthesis2 where the cost of the optimal solution of staying on this singular arc is equal to leave it for hitting the singular arc associated to synthesis1.

Arguments

synthesis1 : the below Synthesis

synthesis2 : the above Synthesis

xmax : the max of the x-axis

Return

t2 : the time

Filtration.__sep — Method

__sep(
    synthesis1::Synthesis,
    synthesis2::Synthesis,
    xmax::Real
) -> SciMLBase.ODESolution

Provide the separation curve of two synthesis by diffetial continuation method.

Arguments

synthesis1 : the below Synthesis

synthesis2 : the above Synthesis

xmax : the max of the x-axis

Return

sol : the separation curve, as an ODESolution

Filtration.classification_plot — Method

classification_plot(
    s::Synthesis;
    SL_nb_points,
    kwargs...
) -> Plots.Plot

Plot a Synthesis, in a classification of optimal control form

Arguments

synthesis : a Synthesis

Keyword arguments (optional)

SLnbpoints : Number of points for the switching locus curve. Initialized to nothing (points of the ODESolution`)

kwargs... : keyword arguments for plot

Returns

plt : the plot

Filtration.dispersal_locus_2 — Method

dispersal_locus_2(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real},
    DL::SciMLBase.ODESolution
) -> SciMLBase.ODESolution

Second dispersal locus curve computed by continuation method. This curve is associated to the singular curve.

A state $[x_1, x_2]$ belongs to the second dispersal curve if $S(x_2, x_1) = 0$, where $S \colon \mathbb R^2 \to \mathbb R$ is defined by

\[S(x_1, x_2) = \varphi_0^+(x_1, x_2) - \varphi_0^{-s+}(x_1, x_2)\]

where $\varphi_0^+$ correponds to the final cost with the control u₊, and where $\varphi_0^{-s+}$ the final cost with the control u₋ ∘ uₛ ∘ u₊.

We suppose that there exsits the switching locus $DL2(x_1)$ such that $S(x_1, DL2(x_1)) = 0$. Since $S$ is constant, we have

\[ \frac{\partial S}{\partial x_1}(x_1, DL2(x_1)) + \frac{\partial S}{\partial x_2}(x_1, DL2(x_1)) DL2'(x_1) = 0\]

By the previous hypothesis, $\frac{\partial S}{\partial x_2}(x_1, DL2(x_1))$ is invertible. We have so

\[DL2'(x_1) = -\left( \frac{\partial S}{\partial x_2}(x_1, DL2(x_1)) \right)^{-1} \frac{\partial S}{\partial x_1}(x_1, DL2(x_1)).\]

The continuation method consists to solve this ODE with the initial condition $DL2(x_s) = x_f - \Delta x_2$. Thanks to the combinaison of ForwardDiff and OrdinaryDiffEq package, the derivative of $\varphi_0^{+}$ and $\varphi_0^{-s+}$ are computed by automatic differentiation methods by using the variational equations.

The end of the dispersal locus curve corresponds to the ylim[2]

Warning

This function must be used only when a second dispersal locus happened, ie. when the end of the first dispersal locus is below ylim[2].

Filtration.get_arrow_time — Method

get_arrow_time(
    value::Vector{<:Real},
    sol::SciMLBase.ODESolution,
    axis::Symbol
) -> Vector{<:Real}

Compute the time $t$ for which sol(t) ∈ y_value

Arguments

y_value : vector of y-axis value

sol : state trajectory as an ODESolution

Return

times : Vector of times

Filtration.real_solve_control — Method

real_solve_control(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real,
    N::Int64
) -> Solution

Provide the real solution, where the control is only composed by -1 and 1 values. In this method, the control is approximated by -1 and 1 with respect to the value of uₛ.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

N : the number of subdivision of the singular arc

Return

real_sol : the Solution

Filtration.real_solve_optimal — Method

real_solve_optimal(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real,
    N::Int64;
    verbose
) -> Solution

Provide the optimal real solution, where the control is only composed by -1 and 1 values. In this method, the global constraints non linear optimization problem is constructed and solved.

Warning

This method must be long when N become large.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

N : the number of subdivision of the singular arc

Keyword arguments (optional)

verbose : if true, print the optimization process (default is false)

Return

real_sol : the Solution

Filtration.real_solve_singular — Method

real_solve_singular(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real,
    N::Int64;
    verbose
) -> Solution

Provide a real solution, where the control is only composed by -1 and 1 values. In this method, the singular arc is subdivided into N intervals. On each subintervals, we ensure that the solution minimize the cost and satisfy that the begining and the end of the state $x\_1$ on each subintervals corresponds to the value of the singular arc $x\_s$.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

N : the number of subdivision of the singular arc

Keyword arguments (optional)

verbose : if true, print the optimization process (default is false)

Return

real_sol : the Solution

Filtration.synthesis_plot — Method

synthesis_plot(
    multisynthesis::MultiSynthesis;
    kwargs...
) -> Plots.Plot

Plot a MultiSynthesis, in a feedback form.

Arguments

synthesis : a MultiSynthesis

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Returns

plt : the plot

Filtration.synthesis_plot — Method

synthesis_plot(
    s::Synthesis;
    SL_nb_points,
    kwargs...
) -> Plots.Plot

Plot a Synthesis, in a feedback form.

Arguments

s : a Synthesis

Keyword arguments (optional)

SLnbpoints : Number of points for the switching locus curve. Initialized to nothing (points of the ODESolution`)

kwargs... : keyword arguments for plot

Returns

plt : the plot