Filtration.jl API

For the developers, here are the private methods.

Index

Filtration.Filtration

Filtration.Control

Filtration.Extremal

Filtration.Model

Filtration.Model

Filtration.Model

Filtration.Solution

Filtration.Synthesis

Filtration.Synthesis

Filtration.add_solution_to_synthesis!

Filtration.check_root

Filtration.check_root

Filtration.compute_phi

Filtration.construct_extremal

Filtration.costate

Filtration.simulate

Filtration.solution

Filtration.solve_MP

Filtration.state

Documentation

Filtration.Filtration — Module

Filtration module.

Lists all the imported modules and packages:

Base

Core

DataInterpolations

DocStringExtensions

FiniteDifferences

ForwardDiff

LaTeXStrings

NonlinearSolve

OptimalControl

OrdinaryDiffEq

Plots

Roots

List of all the exported names:

Control

Extremal

Model

Solution

Synthesis

add_solution_to_synthesis!

check_root

compute_phi

construct_extremal

costate

plot

plot!

simulate

solution

solve_MP

state

to_real_control

Filtration.Control — Type

Structure of a Control, composed by

control : a vector of functions

time : a vector of times

N : the number of intervals

Filtration.Extremal — Type

Structure of a Solution, composed by

z : a vector of ODESolution

control : a Control

time : a vector of times

N : the number of intervals

Filtration.Model — Type

Structure of a Model, composed by:

e₁ : the function $e_1(x_1)$

e₁ : the function $e_2(x_1)$

f₁ : the function $f_1(x_1)$

f₂ : the function $f_2(x_1)$

g₁ : the function $g_1(x_1)$

g₂ : the function $g_2(x_1)$

e₊ : the function $e_+(x_1)$

e₋ : the function $e_-(x_1)$

f₊ : the function $f_+(x_1)$

f₋ : the function $f_-(x_1)$

g₊ : the function $g_+(x_1)$

g₋ : the function $g_-(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

λ₀ : the constant associated to the criteria type

dynamics! : the cost and state dynamics function

de₊ : the function $e_+'(x_1)$, compute thanks to ForwardDiff package

de₋ : the function $e_-'(x_1)$, compute thanks to ForwardDiff package

df₊ : the function $f_+'(x_1)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(x_1)$, compute thanks to ForwardDiff package

dg₊ : the function $g_+'(x_1)$, compute thanks to ForwardDiff package

dg₋ : the function $g_-'(x_1)$, compute thanks to ForwardDiff package

xₛ : the singular state $x_s$

uₛ : the singular control $u_s$

λ₁ₛ : the singular costate $\lambda_{1,s}$ associated to the state $x_1$

λ₂ₛ : the singular costate $\lambda_{2,s}$ associated to the state $x_2$

Δx₂ : the value $\Delta_x = x_f - x_2(T)$, where $T$ is the final time to leave the singular arc

Filtration.Model — Method

Model(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    g₁::Function,
    g₂::Function,
    p::Vector{<:Real},
    criteria_type::Symbol;
    initial_guess
) -> Model

Constructor for Model.

Arguments

e₁ : the function $e_1(x₁)$

e₂ : the function $e_2(x₁)$

f₁ : the function $f_1(x₁)$

f₂ : the function $f_2(x₁)$

g₁ : the function $g_1(x₁)$

g₂ : the function $g_2(x₁)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

Keyword arguments (optional)

initial_guess : initial guess for the singular state $x_s$, initialized to 1.

Returns

model : the Model

Filtration.Model — Method

Model(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    p::Vector{<:Real},
    criteria_type::Symbol;
    initial_guess
) -> Model

Constructor for Model structure.

Note

This function is a wrapper of Model function, where the functions $g_+$ and $g_-$ are set to 0 and 1 functions, respectively. This leads to $\dot x_2 = 1$ with $x_2(t_0) = 0$, ie. $x_2(t) = t$.

Arguments

e₊ : the function $e_+(x_1)$

e₋ : the function $e_-(x_1)$

f₊ : the function $f_+(x_1)$

f₋ : the function $f_-(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max`

Keyword arguments (optional)

initial_guess : initial guess for the singular state $x_s$, initialized to 1.

Returns

model : the Model model

Filtration.Solution — Type

Structure of a Solution, composed by

state : a vector of ODESolution

control : a Control

time : a vector of times

N : the number of intervals

Filtration.Synthesis — Type

Type of a optimal synthesis, composed by:

model : a Model

tlim : the limits of time $t$

xlim : the limits of mass $m$

cut₋ : the cut₋ function

cut₊ : the cut₊ function

SL : the switching locus

DL : the dispersal locus

end_SL : end of the switching locus curve

Filtration.Synthesis — Method

Synthesis(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> Synthesis

Constructor for Synthesis structure.

Arguments

model : a Model

xlim : the limits of x-axis, which corresponds to the $x_2$ state

ylim : the limits of y-axis, which corresponds to the $x_1$ state

Returns

synthesis : the Synthesis

Filtration.add_solution_to_synthesis! — Method

add_solution_to_synthesis!(
    plt::Plots.Plot,
    solution::Solution;
    element,
    axis_arrow,
    arrow,
    color,
    nb_pts,
    kwargs...
)

Add a Solution to a synthesis plot.

Arguments

plt : the synthesis plot

solution : the Solution

Keyword arguments (optional)

element : list of index of solution.state to plot, initialized to [1, ..., solution.N]

arrow : vector of y coordinates to make arrows, initialized to empty

color : color of trajectory, initialized to gray

nb_pts : number of points for each sub-segments, initialized to 5

kwargs... : keyword arguments for plot

Return

plt : the plot

Filtration.check_root — Method

check_root(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    g₁::Function,
    g₂::Function,
    p::Union{Nothing, Vector{<:Real}},
    criteria_type::Symbol;
    xlim,
    nb_pts,
    kwargs...
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the Model type.

Arguments

e₁ : the function $e_1(x_1)$

e₂ : the function $e_2(x_1)$

f₁ : the function $f_1(x_1)$

f₂ : the function $f_2(x_1)$

g₁ : the function $g_1(x_1)$

g₂ : the function $g_2(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

Keyword arguments (optional)

xlim : limits for x axis, initialized to (0,10)

nb_pts: number of points used for the plot, initialized to 100

kwargs... : keyword arguments for plot

Return

plt : the plot

Filtration.check_root — Method

check_root(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    p::Vector{<:Real},
    criteria_type::Symbol;
    xlim,
    nb_pts,
    kwargs...
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the Model type.

Note

This function is a wrapper of check_root function, where the functions $g_+$ and $g_-$ are set to 0 and 1 functions, respectively.

Arguments

e₊ : the function $f_+(x_1)$

e₋ : the function $f_-(x_1)$

f₊ : the function $g_+(x_1)$

f₋ : the function $g_-(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

Keyword arguments (optional)

xlim : limits for x axis, initialized to (0,10)

nb_pts: number of points used for the plot, initialized to 100

kwargs... : keyword arguments for plot

Return

plt : the plot

Filtration.compute_phi — Method

compute_phi(
    t::Real,
    extremal::Extremal,
    model::Model;
    p
) -> Real

Compute the value of the function t \mapsto \phi(x(t), p(t)) on a given time t of an Extremal computed on a Model.

Arguments

t : a real time $t$

extremal : the Extremal

model : the Model

Keyword arguments (optional)

p : a vector of parameter for the model, initialized by the parametes of the model.

Filtration.construct_extremal — Method

construct_extremal(
    model::Model,
    solution::Solution;
    p
) -> Extremal

Construct an Extremal from a Solution from a Model.

Arguments

model : a Model

solution: a Solution

Return

extremal : the Extremal

Keyword arguments (optional)

p : a vector of parameter for the model, initialized by the parametes of the model.

Filtration.costate — Method

costate(
    costate::Extremal
) -> Filtration.var"#132#133"{Extremal}

Compute the state function $t \mapsto p(t)$ from an $Extremal$.

Arguments

extremal : the Extremal

Return

costate : the function $t \mapsto p(t)$

Filtration.simulate — Method

simulate(
    model::Model,
    control::Control,
    x0::Vector{<:Real};
    p
) -> Solution

Simulate a Control on a Model, starting from and initial state $x_0$.

Arguments

model : a Model

control : a Control

x0: and intial state $x_0$

Keyword arguments (optional)

p : a vector of parameter for the model, initialized by the parametes of the model.

Return

solution : the Solution

Filtration.solution — Method

solution(
    synthesis::Synthesis,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Solution

Wrapper to get solution from synthesis.

Arguments

model : a Model

init_x₂ : the initial state $x_1(t_0)$

end_x₂ : the final state $x_1(t_f)$

init_x₁ : the initial state $m_0$

Returns

sol: the Solution

Filtration.solve_MP — Method

solve_MP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real;
    SL
) -> Solution

Solve the optimal control problem associated to a Model model by indirect method, with the structure given by $u_- \circ u_+$, and when the change is done by intersection with the switching curve

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_1(t_0)$

end_x₂ : the final state $x_1(t_f)$

init_x₁ : the initial state $x_2(t_0)$

Returns

sol: the Solution

Filtration.state — Method

state(
    traj::Union{Extremal, Solution}
) -> Filtration.var"#130#131"

Compute the state function $t \mapsto x(t)$ from a Solution or an $Extremal$.

Arguments

traj : the Solution or the Extremal

Return

state : the function $t \mapsto x(t)$

RecipesBase.plot! — Method

plot!(
    plt::Plots.Plot,
    control::Control;
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot a control

Arguments

control : a Control

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    control::Control;
    color,
    ylim,
    label,
    kwargs...
) -> Plots.Plot

Plot a control

Arguments

control : a Control

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    extremal::Extremal;
    size,
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot an extremal

Arguments

extremal : an Extremal

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    solution::Solution;
    size,
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot a solution

Arguments

solution : a Solution

Keyword arguments (optional)

kwargs... : keyword arguments for plot

Return

plt : the plot