Filtration.jl API

For the developers, here are the private methods.

Index

Filtration.Filtration

Filtration.Control

Filtration.Control

Filtration.Costate

Filtration.Costate

Filtration.Extremal

Filtration.Model

Filtration.Model

Filtration.Model

Filtration.MultiModel

Filtration.MultiModel

Filtration.MultiModel

Filtration.MultiSynthesis

Filtration.MultiSynthesis

Filtration.Solution

Filtration.State

Filtration.State

Filtration.Synthesis

Filtration.Synthesis

Filtration.add_solution_to_synthesis!

Filtration.check_root

Filtration.check_root

Filtration.compute_phi

Filtration.construct_extremal

Filtration.control

Filtration.control

Filtration.costate

Filtration.cuts

Filtration.dispersal_locus

Filtration.real_solve

Filtration.simulate

Filtration.solution

Filtration.solve

Filtration.solve_MP

Filtration.solve_MSP

Filtration.solve_P

Filtration.solve_PSP

Filtration.solve_SP

Filtration.state

Filtration.state

Filtration.switching_locus

Documentation

Filtration.Filtration — Module

Filtration module.

Lists all the imported modules and packages:

Base

Core

DiffEqCallbacks

DocStringExtensions

ForwardDiff

Ipopt

LaTeXStrings

Optimization

OptimizationMOI

OptimizationOptimJL

OrdinaryDiffEq

Plots

Roots

List of all the exported names:

Control

Costate

Extremal

Model

MultiModel

MultiSynthesis

Pareto

SimulationModel

Solution

State

Synthesis

add_solution_to_synthesis!

check_root

compute_phi

construct_extremal

control

costate

cuts

dispersal_locus

fit

plot

plot!

real_solve

show

simulate

solution

solve

solve_MP

solve_MSP

solve_P

solve_PSP

solve_SP

state

switching_locus

Filtration.Control — Type

Structure of a Control, composed by

control : a vector of functions

times : a vector of times

N : the number of intervals

m : size of the control

Filtration.Control — Method

Callable method for the Control structure. This return the value of the control at time t.

Arguments

t : the time $t$

Return

u : the value of the control $u(t)$ at the time $t$

Filtration.Costate — Type

Structure for a Costate, composed by

costate : a vector of ODESolution

time : a vector of times

N : the number of intervals

λ0 : the initial costate $λ_0$

n : number of state / co-state

Filtration.Costate — Method

Callable method for the Costate structure. This return the value of the costate at time t.

Arguments

t : the time $t$

Return

x : the value of the costate $\lambda(t)$ at the time $t$

Filtration.Extremal — Type

Structure of a Solution, composed by

state : a State

costate : a Costate

control : a Control

time : a vector of times

N : the number of intervals

Filtration.Model — Type

Structure of a Model, composed by:

e₁ : the function $e_1(x_1)$

e₁ : the function $e_2(x_1)$

f₁ : the function $f_1(x_1)$

f₂ : the function $f_2(x_1)$

g₁ : the function $g_1(x_1)$

g₂ : the function $g_2(x_1)$

e₊ : the function $e_+(x_1)$

e₋ : the function $e_-(x_1)$

f₊ : the function $f_+(x_1)$

f₋ : the function $f_-(x_1)$

g₊ : the function $g_+(x_1)$

g₋ : the function $g_-(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

λ₀ : the constant associated to the criteria type

dynamics! : the cost and state dynamics function

de₊ : the function $e_+'(x_1)$, compute thanks to ForwardDiff package

de₋ : the function $e_-'(x_1)$, compute thanks to ForwardDiff package

df₊ : the function $f_+'(x_1)$, compute thanks to ForwardDiff package

df₋ : the function $f_-'(x_1)$, compute thanks to ForwardDiff package

dg₊ : the function $g_+'(x_1)$, compute thanks to ForwardDiff package

dg₋ : the function $g_-'(x_1)$, compute thanks to ForwardDiff package

xₛ : the singular state $x_s$

uₛ : the singular control $u_s$

λ₁ₛ : the singular costate $\lambda_{1,s}$ associated to the state $x_1$

λ₂ₛ : the singular costate $\lambda_{2,s}$ associated to the state $x_2$

Δx₂ : the value $\Delta_x = x_f - x_2(T)$, where $T$ is the final time to leave the singular arc

Δt : the difference between the final time and the time to leave the singular arc

Filtration.Model — Method

Model(
    e⁺::Function,
    e⁻::Function,
    f₊::Function,
    f₋::Function,
    g⁺::Function,
    g⁻::Function,
    p::Vector{<:Real},
    criteria_type::Symbol;
    initial_guess,
    n,
    m
) -> Model

Constructor for Model.

Arguments

e₁ : the function $e_1(x₁)$

e₂ : the function $e_2(x₁)$

f₁ : the function $f_1(x₁)$

f₂ : the function $f_2(x₁)$

g₁ : the function $g_1(x₁)$

g₂ : the function $g_2(x₁)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

Keyword arguments (optional)

initial_guess : initial guess for the singular state $x_s$, initialized to 1.

Returns

model : the Model

Filtration.Model — Method

Model(
    e₊::Function,
    e₋::Function,
    f₊::Function,
    f₋::Function,
    p::Vector{<:Real},
    criteria_type::Symbol;
    initial_guess,
    n,
    m
) -> Model

Constructor for a Model.

Note

This function is a wrapper of Model function, where the functions $g_1$ and $g_2$ are set to 1. This leads to $\dot x_2 = 1$ with $x_2(t_0) = 0$, ie. $x_2(t) = t$.

Arguments

e₁ : the function $e_1(x_1)$

e₂ : the function $e_2(x_1)$

f₁ : the function $f_1(x_1)$

f₂ : the function $f_2(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max`

Keyword arguments (optional)

initial_guess : initial guess for the singular state $x_s$, initialized to 1.

Returns

model : the Model model

Filtration.MultiModel — Type

Structure of a MultiModel, composed by

models : a vector of Models, which corresponds to turnpike

__models : a vector of Models, which corresponds to anti-turnpike

xₛ : vector of singular state value, which corresponds to turnpike

__xₛ : vector of singular state value, which corresponds to anti-turnpike

N : number of models

__N : numbers of __models

Filtration.MultiModel — Method

MultiModel(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    g₁::Function,
    g₂::Function,
    p::Vector{<:Real},
    criteria_type::Symbol,
    initial_guesses::Vector{<:Real}
) -> MultiModel

Constructor for MultiModel

Arguments

e₁ : the function $e_1(x₁)$

e₂ : the function $e_2(x₁)$

f₁ : the function $f_1(x₁)$

f₂ : the function $f_2(x₁)$

g₁ : the function $g_1(x₁)$

g₂ : the function $g_2(x₁)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

initial_guesses : initial guesses for all singular states $x_s$

Returns

multimodel : the MultiModel

Filtration.MultiModel — Method

MultiModel(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    p::Vector{<:Real},
    criteria_type::Symbol,
    initial_guesses::Vector{<:Real}
) -> MultiModel

Constructor for a MultiModel.

Note

This function is a wrapper of MultiModel function, where the functions $g_1$ and $g_2$ are set to 1. This leads to $\dot x_2 = 1$ with $x_2(t_0) = 0$, ie. $x_2(t) = t$.

Arguments

e₁ : the function $e_1(x_1)$

e₂ : the function $e_2(x_1)$

f₁ : the function $f_1(x_1)$

f₂ : the function $f_2(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max`

initial_guesses : initial guesses for all singular states $x_s$.

Keyword arguments (optional)

Returns

multimodel : the MultiModel

Filtration.MultiSynthesis — Type

Structure of a MultiSynthesis, composed by

synthesis : the vector of Synthesis

sep : a vector of ODESolution that connect synthesis

N : number of Synthesis

Filtration.MultiSynthesis — Method

MultiSynthesis(
    multimodel::MultiModel,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> MultiSynthesis

Constructor for MultiSynthesis

Arguments

multimodel : a MultiModel

xlim : limits of x-axis

ylim : limits of y-axis

Return

multisynthesis : a MultiSynthesis

Filtration.Solution — Type

Structure of a Solution, composed by

state : a State

control : a Control

time : a vector of times

N : the number of intervals

x0 : the initial state $x(t_0) = x_0$

Filtration.State — Type

Structure of a State, composed by

state : a vector of ODESolution

times : a vector of times

N : the number of intervals

n : number of states

Filtration.State — Method

Callable method for the State structure. This return the value of the state at time t.

Arguments

t : the time $t$

Return

x : the value of the state $x(t)$ at the time $t$

Filtration.Synthesis — Type

Structure of a optimal synthesis, composed by:

model : a Model

xlim : the limits of the x-axis

ylim : the limits of the y-axis

cut₋ : the cut₋ function

cut₊ : the cut₊ function

SL : the switching locus

DL : the dispersal locus with the switching locus

DL2 : the dispersal locus with the singular curve

Filtration.Synthesis — Method

Synthesis(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> Synthesis

Constructor of a Synthesis.

Arguments

model : a Model

xlim : the limits of x-axis

ylim : the limits of y-axis

Returns

synthesis : the Synthesis

Base.show — Method

show(io::IO, z::Control)

Print a Control

Base.show — Method

show(io::IO, z::Costate)

Print a Costate

Base.show — Method

show(io::IO, z::Extremal)

Print an Extremal

Base.show — Method

show(io::IO, z::Model)

Print a Model

Base.show — Method

show(io::IO, z::MultiModel)

Print a MultiModel

Base.show — Method

show(io::IO, z::MultiSynthesis)

Print a MultiSynthesis

Base.show — Method

show(io::IO, z::Pareto)

Print a Pareto

Base.show — Method

show(io::IO, z::SimulationModel)

Print a Model

Base.show — Method

show(io::IO, z::Solution)

Print a Solution

Base.show — Method

show(io::IO, z::State)

Print a State

Base.show — Method

show(io::IO, z::Synthesis)

Print a Synthesis

Filtration.add_solution_to_synthesis! — Method

add_solution_to_synthesis!(
    plt::Plots.Plot,
    solution::Solution;
    element,
    axis_arrow,
    arrow,
    color,
    nb_pts,
    label,
    kwargs...
) -> Plots.Plot

Add a Solution to a synthesis plot.

Arguments

plt : the synthesis plot

solution : the Solution

Keyword arguments (optional)

element : list of index of solution.state to plot, initialized to [1, ..., solution.N]

-axis_arrow: symbol to indicate the axis for the arrow, initialized to nothing (no arrow). This symbol must be :x or :y.

arrow : vector of axis_arrow coordinates to make arrows, initialized to empty

color : color of trajectory, initialized to gray

nb_pts : number of points for each sub-segments, initialized to 5

kwargs... : keyword arguments for plot

Return

plt : the plot

Filtration.check_root — Method

check_root(
    e⁺::Function,
    e⁻::Function,
    f₊::Function,
    f₋::Function,
    g⁺::Function,
    g⁻::Function,
    p::Union{Nothing, Vector{<:Real}},
    criteria_type::Symbol;
    xlim,
    initial_guess,
    kwargs...
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the Model type.

Arguments

e₁ : the function $e_1(x_1)$

e₂ : the function $e_2(x_1)$

f₁ : the function $f_1(x_1)$

f₂ : the function $f_2(x_1)$

g₁ : the function $g_1(x_1)$

g₂ : the function $g_2(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

Keyword arguments (optional)

xlim : limits for x axis, initialized to (0,10)

nb_pts : number of points used for the plot, initialized to 100

kwargs... : keyword arguments for plot

Return

plt : the plot

Filtration.check_root — Method

check_root(
    e₁::Function,
    e₂::Function,
    f₁::Function,
    f₂::Function,
    p::Vector{<:Real},
    criteria_type::Symbol;
    xlim,
    nb_pts,
    initial_guess,
    kwargs...
) -> Plots.Plot

Plot a function where we need to ensure that there is only one zero of it, on the domain of interest. An initial guess of this zero can be provided to the construtor of the Model type.

Note

This function is a wrapper of check_root function, where the functions $g_1$ and $g_2$ are set 1. This leads to $\dot x_2 = 1$ with $x_2(t_0) = 0$, ie. $x_2(t) = t$.

Arguments

e₁ : the function $f_1(x_1)$

e₂ : the function $f_2(x_1)$

f₁ : the function $g_1(x_1)$

f₂ : the function $g_2(x_1)$

p : parameters of previous functions

criteria_type : the criteria type, either :min or :max

Keyword arguments (optional)

xlim : limits for x axis, initialized to (0,10)

nb_pts: number of points used for the plot, initialized to 100

initial_guess : initial guess(es) for zero of the function. Initialized to nohting.

kwargs... : keyword arguments for plot

Return

plt : the plot

Filtration.compute_phi — Method

compute_phi(
    extremal::Extremal,
    model::Model;
    p
) -> Filtration.var"#phi#106"{Vector{var"#s37"}, Extremal, Model} where var"#s37"<:Real

Compute the function $t \mapsto \phi(x(t), p(t))$ of an Extremal computed on a Model.

Arguments

extremal : the Extremal

model : the Model

Keyword arguments (optional)

p : a vector of parameter for the model (default is the parameters of the model).

Return

phi : the function $\phi$.

Filtration.construct_extremal — Method

construct_extremal(
    model::Model,
    solution::Solution;
    p
) -> Extremal

Construct an Extremal from a Solution from a Model.

Arguments

model : a Model

solution: a Solution

Keyword arguments (optional)

p : a vector of parameter for the model (default is the parametes of the model).

Return

extremal : the Extremal

Filtration.control — Method

control(extremal::Extremal) -> Control

Getter for the Control of an Extremal.

Arguments

extremal : an Extremal

Return

control : the Control of the Extremal

Filtration.control — Method

control(solution::Solution) -> Control

Getter for the Control of a Solution.

Arguments

solution : a Solution

Return

state : the Control of the Solution

Filtration.costate — Method

costate(extremal::Extremal) -> Costate

Getter for the Costate of an Extremal.

Arguments

extremal : an Extremal

Return

costate : the Costate of the Extremal

Filtration.cuts — Method

cuts(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> Tuple{SciMLBase.ODESolution, SciMLBase.ODESolution}

Compute the cuts $c_-(t)$ and $c_+(t)$, which corresponds to the curves which leads to the singular state $x_s$ at time $T$. The function $c₋$ is the solution on $[t_0, T]$ of

\[\left\{ \begin{array}{ll} \dot x_0 = e_-(x_1) - e_+(x_1), & \quad x_0(T) = 0, \\\dot x_1 = f_-(x_1) - f_+(x_1), & \quad x_1(T) = x_s, \\\dot x_2 = g_-(x_1) - g_+(x_1), & \quad x_2(T) = x_f- \Delta_x, \end{array} \right.\]

and the function $m₊$ is the solution on $[T, t_f]$ of

\[\left\{\begin{array}{ll} \dot x_0 = e_-(x_1) + e_+(x_1), & \quad x_O(T) = 0, \\\dot x_1 = f_-(x_1) + f_+(x_1), & \quad x_1(T) = x_s, \\\dot x_2 = g_-(x_1) + g_+(x_1), & \quad x_2(T) = x_f - \Delta_x, \end{array} \right.\]

Warning

The cost $x_0(T)$ at time $T$ is initialized to 0.

Arguments

model : a Model

xlim : the limits of state $x_2$

ylim : the limits of state $x_1$

Returns

cut₋ : the function $c_-$

cut₊ : the function $c_+$

Filtration.dispersal_locus — Method

dispersal_locus(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real},
    SL::SciMLBase.ODESolution,
    cut₋::SciMLBase.ODESolution
) -> SciMLBase.ODESolution

First dispersal locus curve computed by continuation method. This curve is associated to the switching locus curve.

A state $[x_1, x_2]$ belongs to the first dispersal curve if $S(x_2, x_1) = 0$, where $S \colon \mathbb R^2 \to \mathbb R$ is defined by

\[S(x_1, x_2) = \varphi_0^+(x_1, x_2) - \varphi_0^{-+}(x_1, x_2)\]

where $\varphi_0^+$ correponds to the final cost with the control u₊, and where $\varphi_0^{-+}$ the final cost with the control u₋ ∘ u₊.

We suppose that there exsits the switching locus $DL(x_1)$ such that $S(x_1, DL(x_1)) = 0$. Since $S$ is constant, we have

\[ \frac{\partial S}{\partial x_1}(x_1, DL(x_1)) + \frac{\partial S}{\partial x_2}(x_1, DL(x_1)) DL'(x_1) = 0\]

By the previous hypothesis, $\frac{\partial S}{\partial x_2}(x_1, DL(x_1))$ is invertible. We have so

\[DL'(x_1) = -\left( \frac{\partial S}{\partial x_2}(x_1, DL(x_1)) \right)^{-1} \frac{\partial S}{\partial x_1}(x_1, DL(x_1)).\]

The continuation method consists to solve this ODE with the initial condition $DL(x_s) = x_f - \Delta x_2$. Thanks to the combinaison of ForwardDiff and OrdinaryDiffEq package, the derivative of $\varphi_0^{+}$ and $\varphi_0^{-+}$ are computed by automatic differentiation methods by using the variational equations.

The end of the dispersal locus curve corresponds to the ylim[2] bouds or when the dispersal locus curve hit the cut₋ curve.

Warning

This function must be used only when a dispersal locus happened, ie. when the end of the switching curve is below ylim[2].

Filtration.real_solve — Method

real_solve(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real,
    N::Int64;
    method,
    verbose
) -> Solution

Provide a real solution, where the control is only composed by -1 and 1 values, by the method. method can be :control, :state or :optimal.

Warning

If method = :optimal, the computation must be long when N become large.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

N : the number of subdivision of the singular arc

Keyword arguments (optional)

method : method used to get the real solution. Can be :control, :state or :optimal. Initialized to :state

verbose : if true, print the optimization process (default is false)

Return

real_sol : the Solution

Filtration.simulate — Method

simulate(
    model::Model,
    control::Control,
    x0::Vector{<:Real};
    p
) -> Solution

Simulate a Control on a Model, starting from and initial state $x(t_0) = x_0$.

Arguments

model : a Model

control : a Control

x0: and intial state $x(t_0) = x_0$

Keyword arguments (optional)

p : a vector of parameter for the model, (default is the parametes of the model).

Return

solution : the Solution

Filtration.solution — Method

solution(
    synthesis::Synthesis,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Solution

Wrapper to get solution from synthesis.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.solve — Method

solve(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real;
    cut₋,
    cut₊,
    SL,
    DL,
    DL2
) -> Solution

Solve the optimal control problem associated to a Model model by indirect shooting.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.solve_MP — Method

solve_MP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real;
    SL
) -> Solution

Solve the optimal control problem associated to a Model model by indirect method, with the structure given by $u_- \circ u_+$, and when the change is done by intersection with the switching curve

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.solve_MSP — Method

solve_MSP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Solution

Solve the optimal control problem associated to a Model model, with the structure given by $u_- \circ u_s \circ u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.solve_P — Method

solve_P(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Solution

Solve the optimal control problem associated to a Model model, with the structure given by $u_+$.

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.solve_PSP — Method

solve_PSP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Solution

Solve the optimal control problem associated to the membrane filtration model by indirect shooting, with the structure given by $u_+ \circ u_s \circ u_+$

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.solve_SP — Method

solve_SP(
    model::Model,
    init_x₂::Real,
    end_x₂::Real,
    init_x₁::Real
) -> Solution

Solve the optimal control problem associated to the membrane filtration model by indirect shooting, with the structure given by $u_+ \circ u_s \circ u_+$

Warning

This function must be used only if we know the structure. If you don't know the structure of the solution, use solve function instead.

Arguments

model : a Model

init_x₂ : the initial state $x_2(t_0)$

end_x₂ : the final state $x_2(t_f)$

init_x₁ : the initial state $x_1(t_0)$

Returns

sol: the Solution

Filtration.state — Method

state(extremal::Extremal) -> State

Getter for the State of an Extremal.

Arguments

extremal : an Extremal

Return

state : the State of the Extremal

Filtration.state — Method

state(solution::Solution) -> State

Getter for the State of a Solution.

Arguments

solution : a Solution

Return

state : the State of the Solution

Filtration.switching_locus — Method

switching_locus(
    model::Model,
    xlim::Tuple{Real, Real},
    ylim::Tuple{Real, Real}
) -> SciMLBase.ODESolution

Switching locus function computed by continuation method.

A state $[x_1, x_2]$ belongs to the switching curve if $S(x_2, x_1) = 0$, where $S \colon \mathbb R^2 \to \mathbb R$ is defined by

\[S(x_2, x_1) = H(x_1, \varphi_1(x_1, x_2))\]

The function $\varphi_1$ coresponds to the solution at the final time of the state $x_1$ by applying the control $u_+$, and the function $H$ is defined by

\[ H(x_1, x_{1,f}) = \lambda_1(x_1, x_{1,f}) (f_+(x_1) + f_-(x_1)) + \lambda_2(x_{1,f})(g_+(x_1) + g_-(x_1)) - (e_+(x_1) + e_-(x_1))\]

where

\[ \lambda_1(x_1, x_{1,f}) = \frac{e_+(x_1) - \lambda_2(x_{1,f}) g_+(x_1)}{f_+(x_1)}\]

and

\[ \lambda_2(x_{1,f}) = \frac{e_+(x_1) + e_-(x_1)}{g_+(x_1) + g_-(x_1)}. \]

We suppose that there exsits the switching locus $SL(x_1)$ such that $S(x_1, SL(x_1)) = 0$. Since $S$ is constant, we have

\[ \frac{\partial S}{\partial x_1}(x_1, SL(x_1)) + \frac{\partial S}{\partial x_2}(x_1, SL(x_1)) SL'(x_1) = 0\]

By the previous hypothesis, $\frac{\partial S}{\partial x_2}(x_1, SL(x_1))$ is invertible. We have so

\[SL'(x_1) = -\left( \frac{\partial S}{\partial x_2}(x_1, SL(x_1)) \right)^{-1} \frac{\partial S}{\partial x_1}(x_1, SL(x_1)).\]

The continuation method consists to solve this ODE with the initial condition $SL(x_s) = x_f - \Delta x_2$. Thanks to the combinaison of ForwardDiff and OrdinaryDiffEq package, the derivative of $\varphi_1$ is computed by automatic differentiation methods by using the variatomal equations.

Arguments

model: a Model

xlim : the limits of the state $x_2$

ylim : the limits of the state $x_1$

Return

SL : the switching locus function, as an ODESolution

RecipesBase.plot! — Method

plot!(
    plt::Plots.Plot,
    control::Control;
    color,
    label,
    kwargs...
)

Add a Control to a plot

Arguments

plt : a Plot

control : a Control

Keyword arguments (optional)

color : the color of the plot (default is the first color of the classical theme palette)

label : the label of the plot (default is an empy string)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot! — Method

plot!(
    plt::Plots.Plot,
    costate::Costate;
    color,
    label,
    subplot,
    kwargs...
)

Plot a costate

Arguments

costate : a Costate

Keyword arguments (optional)

color : the color of the plot (default is the first color of the classical theme palette)

label : the label of the plot (default is an empy string)

subplot : the initial subplot number (default is 1)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot! — Method

plot!(
    plt::Plots.Plot,
    extremal::Extremal;
    color,
    label,
    kwargs...
)

Add an Extremal on a Plot

Arguments

plt : a Plots.Plot

extremal : an Extremal

Keyword arguments (optional)

color : the color of the plot (default is the first color of the theme palette)

label : the label of the plot (default is "")

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot! — Method

plot!(
    plt::Plots.Plot,
    solution::Solution;
    color,
    kwargs...
) -> Plots.Plot

Add a Solution on a plot

Arguments

plt : a Plots.Plot

solution : a Solution

Keyword arguments (optional)

color : the color of the plot (default is the first color of the theme palette)

label : the label of the plot (default is "")

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot! — Method

plot!(
    plt::Plots.Plot,
    state::State;
    color,
    subplot,
    label,
    kwargs...
)

Add a State to a plot

Arguments

state : a State

Keyword arguments (optional)

color : the color of the plot (default is the first color of the classical theme palette)

label : the label of the plot (default is an empy string)

subplot : the initial subplot number (default is 1)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    control::Control;
    color,
    ylim,
    label,
    titlefontsize,
    kwargs...
) -> Plots.Plot

Plot a Control

Arguments

control : a Control

Keyword arguments (optional)

color : the color of the plot (default is the first color of the classical theme palette)

ylim : the y limits of the plot (default is (-1.1, 1.1))

label : the label of the plot (default is an empy string)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    costate::Costate;
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot a Costate

Arguments

costate : a Costate

Keyword arguments (optional)

color : the color of the plot (default is the first color of the classical theme palette)

label : the label of the plot (default is an empy string)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    extremal::Extremal;
    size,
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot an Extremal

Arguments

extremal : an Extremal

Keyword arguments (optional)

size : the size of the plot (default is (600, 350))

color : the color of the plot (default is the first color of the classical theme palette)

label : the label of the plot (default is empty string)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    s::MultiSynthesis;
    feedback_form,
    kwargs...
) -> Plots.Plot

Plot a MultiSynthesis, in a classification of optimal control form or in a feedback form.

Arguments

multisynthesis : a MultiSynthesis

Keyword arguments (optional)

feedback_form : if true, plot in a feedback form. If false, in a classification of optimal control form. Initialized to true

kwargs... : keyword arguments for plot

Returns

plt : the plot

RecipesBase.plot — Method

plot(
    solution::Solution;
    state_ylabels,
    size,
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot a Solution

Arguments

solution : a Solution

Keyword arguments (optional)

size : the size of the plot (default is (600, 350))

color : the color of the plot (default is the first color of the theme palette)

label : the label of the plot (default is "")

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    state::State;
    ylabels,
    color,
    label,
    kwargs...
) -> Plots.Plot

Plot a State

Arguments

state : a State

Keyword arguments (optional)

color : the color of the plot (default is the first color of the classical theme palette)

label : the label of the plot (default is an empy string)

kwargs... : keyword arguments for plot

Return

plt : the plot

RecipesBase.plot — Method

plot(
    s::Synthesis;
    feedback_form,
    SL_nb_points,
    kwargs...
) -> Plots.Plot

Plot a Synthesis, in a classification of optimal control form or in a feedback form.

Arguments

s : a Synthesis

Keyword arguments (optional)

feedback_form : if true, plot in a feedback form. If false, in a classification of optimal control form. Initialized to false

SLnbpoints : Number of points for the switching locus curve. Initialized to nothing (points of the ODESolution`)

kwargs... : keyword arguments for plot

Returns

plt : the plot