Filtration.jl private functions

Index

Filtration.indirect_solve_MP

Filtration.indirect_solve_MSP

Filtration.indirect_solve_P

Filtration.indirect_solve_PSP

Documentation

Filtration.indirect_solve_MP — Method

indirect_solve_MP(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real,
    tlim::Tuple{Real, Real},
    mlim::Tuple{Real, Real}
) -> Any

Solve the optimal control problem associated to the membrane filtration model by indirect shooting, with the structure given by $u_- \circ u_+$.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial state $m(t_0)$

tlim : tuple of real which corresponds to the limits of time

mlim : tuple of real which corresponds to the limits of mass

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_MSP — Method

indirect_solve_MSP(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to the membrane filtration model with the structure given by $u_- \circ u_s \circ u_+$.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial state $m(t_0)$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_P — Method

indirect_solve_P(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to the membrane filtration model with the structure given by $u_+$.

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial state $m(t_0)$

Returns

sol: the solution, as an ODESolution

Filtration.indirect_solve_PSP — Method

indirect_solve_PSP(
    model::MaxVolume,
    t0::Real,
    tf::Real,
    m0::Real
) -> Any

Solve the optimal control problem associated to the membrane filtration model by indirect shooting, with the structure given by $u_+ \circ u_s \circ u_+$

Arguments

model : a MaxVolume

t0 : the initial time $t_0$

tf : the final time $t_f$

m0 : the initial state $m(t_0)$

Returns

sol: the solution, as an ODESolution